+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Fix `--regex-pattern` being ignored by the CLI.
 - Support Python 3.14.
 - Drop support for Python 3.9 and lower.
 - Use tox for local test runs and in CI.